Reorganization.

+++ 0_vanilla/path_planner.py
@@ -2,7 +2,7 @@ import numpy as np
 
 
 def terrain_value_iteration(terrain_mtx, target, max_horizon, append=1):
-    terrain_mtx[target[0], target[1]] = 0.0 # set cost at the target to 0
+    terrain_mtx[target[0], target[1]] = 0 # set cost at the target to 0
 
     nX = terrain_mtx.shape[0]
     nY = terrain_mtx.shape[1]
@@ -18,7 +18,7 @@ def terrain_value_iteration(terrain_mtx, target, max_horizon, append=1):
     # Create J matrix
     # row -> x
     # column -> y
-    J = np.zeros((nX, nY))
+    J = np.zeros((nX, nY), dtype=np.int32)
 
     u = np.array([[0,0], [1,0], [0, 1], [-1,0], [0,-1], [-1,-1], [1, 1], [-1,1], [1,-1]], dtype=np.int32)
